Moved doc stuff from game.py/json to meta.py/json
