+++ src/__init__.py
@@ -9,7 +9,8 @@ from worlds.generic.Rules import forbid_items_for_player
 from worlds.LauncherComponents import Component, SuffixIdentifier, components, Type, launch_subprocess
 
 from .Data import item_table, location_table, region_table, category_table
-from .Game import game_name, filler_item_name, starting_items, set_world_doc, set_world_webworld
+from .Game import game_name, filler_item_name, starting_items
+from .Meta import set_world_doc, set_world_webworld
 from .Locations import location_id_to_name, location_name_to_id, location_name_to_location, location_name_groups
 from .Items import item_id_to_name, item_name_to_id, item_name_to_item, item_name_groups
 from .DataValidation import runGenerationDataValidation
