+++ src/data/meta.json
@@ -1,0 +1,27 @@
+{
+    "_Comment": "remove the _ before a non-comment tag to enable it",
+    "Docs": {
+        "_For more info check": "for more info check https://github.com/ArchipelagoMW/Archipelago/blob/main/docs/world%20api.md#docstrings",
+        "apworld_description": [
+            "Manual games allow you to set custom check locations and custom item names that will be rolled into a multiworld.",
+            "This allows any variety of game -- PC, console, board games, Microsoft Word memes... really anything -- to be part of a multiworld randomizer.",
+            "The key component to including these games is some level of manual restriction. Since the items are not actually withheld from the player,",
+            "the player must manually refrain from using these gathered items until the tracker shows that they have been acquired or sent."
+        ],
+        "web": {
+            "_Comment": "theres a lot more you can do in the web section.",
+            "_for more info check": "https://github.com/ArchipelagoMW/Archipelago/blob/main/docs/world%20api.md#webworld-class",
+            "tutorials": [
+                {
+                    "name": "Multiworld Setup Guide",
+                    "description": "A guide to setting up manual game integration for Archipelago multiworld games.",
+                    "language": "English",
+                    "file_name": "setup_en.md",
+                    "link": "setup/en",
+                    "_authors": ["Fuzzy"],
+                    "_comment": "if you dont include the authors it will use creator from game.json"
+                }
+            ]
+        }
+    }
+}
